// src/generated/server/alsaglobal/intelligence/v1/service_server.ts
export interface SecurityAdvisoryItem {
  title: string;
  link: string;
  pubDate: string;
  source: string;
  sourceCountry: string;
  level: string;
  country: string;
}